feat: create Experience component and refactor experience display in index page
